model-server/requirements.txt
accelerate==0.27.2
torch==2.2.1
fastapi==0.110.0
pyannote-audio==3.1.1
transformers==4.38.2
numpy==1.26.4
torchaudio==2.2.1
uvicorn==0.27.1
httpx==0.27.0
python-multipart==0.0.9
pydantic-settings==2.2.1
pytest==8.1.1